Add files via upload
*** THIS CODE IS INTENDED FOR EDUCATIONAL PURPOSES ONLY ***

It's import to understand that there are five penetration testing phases: reconnaissance, scanning, vulnerability assessment, exploitation, and reporting. 


The provided python script is part of the "reconnaissance" phase in the hacking or pen testing lifecycle process. The reconnaissance phase is the hacking lifecycle and involves gathering information about the target system or network. In this case, the code is performing a type of reconnaissance known as "port scanning."

Port scanning is a technique used to discover open ports on a target system. Open ports can be potential entry points for attackers, and knowledge of open ports helps in understanding the services running on the target system. The code is attempting to identify open ports on the specified target(s) and, in some cases, retrieve banner information to gain additional details about the services running on those ports.

It's important to note that while understanding and securing systems is a legitimate and necessary part of cybersecurity, using such tools and techniques for unauthorized purposes is illegal and unethical. Always ensure that you have explicit permission before conducting any security testing or scanning activities on systems that you do not own or have explicit permission to test.